refactor: `layout-default` –> `layout`
